Distinguish first button from the Group one
For CHACON DIO remote, the G button (the ALL one ) is sent with the same
device id as the first button, but with ALL_ON or ALL_OFF a command instead
of ON or OFF.

Append G to this first button id when the command is ALL_ON or ALL_OFF to
uniquely identify the G button.

Fixes #10.